Split a service into 2 separate responsibilities, to remove a dependency cycle.
